work on history feature

+++ ApiDoc/Models/VersionedItem.cs
@@ -13,7 +13,7 @@ namespace ApiDoc.Models
         public DateTime ChangeDate { get; set; }
         [Required]
         public string ChangeNote { get; set; }
-        public long RevisionNumber { get; set; }
+        public int RevisionNumber { get; set; }
         public bool Deleted { get; set; }
 
         public abstract string GetWikiUrlString();
